driver: consolidate creating threads
main thread calls rdcyc() that reads the cpu-specific timestamp counter
for software timing loops to control the clock going to the tubes, so
the main thread is restricted to cpu #1 so the same timestamp counter is
used throughout.

the other threads can use any cpu, so use pthread attributes to create
them with all cpus enabled.

this only applies when driving the actual tubes or the zynq simulator.
the other simulators (nohw, csrc) don't use rdcyc() for timing so all
threads can run on all cpus.